requirements.txt
async-timeout==4.0.3
attrs==23.1.0
autobahn==19.11.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.5
exceptiongroup==1.1.3
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kiteconnect==5.0.0
numpy==1.26.1
outcome==1.3.0.post0
pandas==2.1.2
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
pyotp==2.9.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
redis==5.0.1
requests==2.31.0
selenium==4.15.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wsproto==1.2.0
zope.interface==6.1